requirements.txt
flask>=3.0,<4.0
flask-cors>=4.0,<5.0
gunicorn>=21.0,<23.0
google-auth>=2.20,<3.0
google-auth-oauthlib>=1.0,<2.0
google-api-python-client>=2.90,<3.0
